Improve qsort_r detection on a future version of FreeBSD.
The Austin Group is standardizing qsort_r(3) in an upcoming revision
of the standard, and FreeBSD will adopt the standarized qsort_r(3)
interface in future releases, and source level compatibility will
be maintained by using C _Generic, which will be available as a
macro called qsort_r.

Detect this on FreeBSD by checking the existence of qsort_r,
and use the GNU function signature, which will be a direct
call of qsort_r and avoids the translations.